HTTP v1 deprecation
