+++ component/http/check_test.go
@@ -1,4 +1,4 @@
-package v2
+package http
 
 import (
 	"net/http"
