+++ component/http/component_option.go
@@ -1,4 +1,4 @@
-package v2
+package http
 
 import (
 	"errors"
