+++ component/http/observability_test.go
@@ -1,4 +1,4 @@
-package v2
+package http
 
 import (
 	"fmt"
