+++ examples/service/http.go
@@ -6,8 +6,8 @@ import (
 	"net/http"
 
 	"github.com/beatlabs/patron"
-	v2 "github.com/beatlabs/patron/component/http/v2"
-	"github.com/beatlabs/patron/component/http/v2/router/httprouter"
+	patronhttp "github.com/beatlabs/patron/component/http"
+	"github.com/beatlabs/patron/component/http/router/httprouter"
 	"github.com/beatlabs/patron/log"
 )
 
@@ -25,8 +25,8 @@ func createHttpRouter() (patron.Component, error) {
 		rw.WriteHeader(http.StatusOK)
 	}
 
-	var routes v2.Routes
-	routes.Append(v2.NewGetRoute("/", handler))
+	var routes patronhttp.Routes
+	routes.Append(patronhttp.NewGetRoute("/", handler))
 	rr, err := routes.Result()
 	if err != nil {
 		return nil, fmt.Errorf("failed to create routes: %w", err)
@@ -37,5 +37,5 @@ func createHttpRouter() (patron.Component, error) {
 		return nil, fmt.Errorf("failed to create http router: %w", err)
 	}
 
-	return v2.New(router)
+	return patronhttp.New(router)
 }
